Split the implementation helpers out of the test suite
